refactor(community): components/ → ui/ 폴더 네이밍 통일
- components/ 폴더를 ui/로 이름 변경 (도메인 구조 규칙 통일)
- sections/ 내부 import 경로 수정
