+++ apps/web/src/domain/community/ui/ChatPreview.tsx
