+++ apps/web/src/domain/community/ui/JobCategoryCard.tsx
