+++ apps/web/src/domain/community/ui/QnaChatCard.tsx
